feat: AdminTweets.vue delete function
